Move APIs back to SolutionRestoreManager namespace
