+++ src/NuGet.Clients/NuGet.VisualStudio/NuGet.VisualStudio.csproj
@@ -43,10 +43,6 @@
     </EmbeddedResource>
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="SolutionRestoreManager\" />
-  </ItemGroup>
-
   <Import Project="$(BuildCommonDirectory)common.targets" />
   <Import Project="Sdk.targets" Sdk="Microsoft.NET.Sdk" />
 </Project>
