+++ src/NuGet.Clients/NuGet.VisualStudio/PublicAPI.Unshipped.txt
@@ -1,6 +1,6 @@
-NuGet.VisualStudio.IVsProjectRestoreInfoSource
-NuGet.VisualStudio.IVsProjectRestoreInfoSource.HasPendingNomination.get -> bool
-NuGet.VisualStudio.IVsProjectRestoreInfoSource.Name.get -> string
-NuGet.VisualStudio.IVsProjectRestoreInfoSource.WhenNominated(System.Threading.CancellationToken cancellationToken) -> System.Threading.Tasks.Task
-NuGet.VisualStudio.IVsSolutionRestoreService4
-NuGet.VisualStudio.IVsSolutionRestoreService4.RegisterRestoreInfoSourceAsync(NuGet.VisualStudio.IVsProjectRestoreInfoSource restoreInfoSource, System.Threading.CancellationToken cancellationToken) -> System.Threading.Tasks.Task
+NuGet.SolutionRestoreManager.IVsProjectRestoreInfoSource
+NuGet.SolutionRestoreManager.IVsProjectRestoreInfoSource.HasPendingNomination.get -> bool
+NuGet.SolutionRestoreManager.IVsProjectRestoreInfoSource.Name.get -> string
+NuGet.SolutionRestoreManager.IVsProjectRestoreInfoSource.WhenNominated(System.Threading.CancellationToken cancellationToken) -> System.Threading.Tasks.Task
+NuGet.SolutionRestoreManager.IVsSolutionRestoreService4
+NuGet.SolutionRestoreManager.IVsSolutionRestoreService4.RegisterRestoreInfoSourceAsync(NuGet.SolutionRestoreManager.IVsProjectRestoreInfoSource restoreInfoSource, System.Threading.CancellationToken cancellationToken) -> System.Threading.Tasks.Task
