+++ src/NuGet.Clients/NuGet.VisualStudio/SolutionRestoreManager/IVsReferenceItems.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections;
 using System.Runtime.InteropServices;
 
-namespace NuGet.VisualStudio
+namespace NuGet.SolutionRestoreManager
 {
     /// <summary>
     /// Collection of restore reference items.
