+++ src/NuGet.Clients/NuGet.VisualStudio/SolutionRestoreManager/IVsSolutionRestoreStatusProvider.cs
@@ -7,7 +7,7 @@ using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace NuGet.VisualStudio
+namespace NuGet.SolutionRestoreManager
 {
     /// <summary>
     /// Provides the status of IVsSolutionRestore.
@@ -19,7 +19,7 @@ namespace NuGet.VisualStudio
         /// <summary>
         /// IsRestoreCompleteAsync indicates whether or not automatic package restore has pending work.
         /// Automatic package restore applies for both packages.config and PackageReference projects.
-        /// 
+        ///
         /// Returns true if all projects in the solution that require nomination have been nominated for restore and all pending restores have completed.
         /// The result does not indicate that restore completed successfully, a failed restore will still return true.
         /// </summary>
