+++ src/NuGet.Clients/NuGet.VisualStudio/SolutionRestoreManager/IVsTargetFrameworks2.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections;
 using System.Runtime.InteropServices;
 
-namespace NuGet.VisualStudio
+namespace NuGet.SolutionRestoreManager
 {
     /// <summary>
     /// Represents a collection of target framework metadata items
