+++ test/NuGet.Clients.Tests/NuGet.SolutionRestoreManager.Test/VsProjectProperty.cs
@@ -2,7 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using NuGet.VisualStudio;
 
 namespace NuGet.SolutionRestoreManager.Test
 {
