+++ test/NuGet.Clients.Tests/NuGet.SolutionRestoreManager.Test/VsReferenceItems.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using NuGet.VisualStudio;
 
 namespace NuGet.SolutionRestoreManager.Test
 {
