+++ test/NuGet.Clients.Tests/NuGet.SolutionRestoreManager.Test/VsTargetFrameworks.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Collections.Generic;
-using NuGet.VisualStudio;
 
 namespace NuGet.SolutionRestoreManager.Test
 {
